Start cleaning newly added path functions

+++ source/D2Common/include/Path/PathMisc.h
@@ -93,7 +93,7 @@ void __stdcall D2Common_10233(D2DynamicPathStrc* pDynamicPath);
 //D2Common.0x6FDADF00 (#10230)
 BOOL __stdcall D2Common_10230(D2DynamicPathStrc* pDynamicPath, int a2, D2RoomStrc* pRoom, unsigned int a4, __int16 a5);
 //D2Common.0x6FDADC20 (#10231)
-signed int __stdcall D2Common_10231(D2DynamicPathStrc * a1, D2UnitStrc * a2, D2RoomStrc * pRooms, int nX, int nY);
+signed int __stdcall D2Common_10231(D2DynamicPathStrc * a1, D2UnitStrc * a2, D2RoomStrc * pRooms, uint16_t nX, uint16_t nY);
 
 BOOL __stdcall D2Common_10234(D2DynamicPathStrc* pDynamicPath);
 //D2Common.0x6FDAE520 (#10235)
